improves moyennes and notes to get nomMatiere from BDD, changes type of Note from int to Double

+++ src/main/java/fr/utln/atlas/projethyp/entities/Devoir.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class Devoir implements Entity{
     private int id;
     private int idUtilisateur;
-    private int note;
+    private Double note;
     private String commentaire;
 
     public enum TypeDevoir {
